requirements/base.txt
BeautifulSoup==3.2.1
Django==1.6.5
Pillow==2.5.3
South==1.0
beautifulsoup4==4.3.2
django-ace==1.0.1
django-analytics==0.0.1
django-appconf==0.6
django-blog-zinnia==0.14.2
django-braces==1.2.2
django-classy-tags==0.5.1
django-contrib-comments==1.5
django-dbtemplates==1.3
django-debug-toolbar==1.2.1
django-endless-pagination==2.0
django-filer==0.9.7
django-geckoboard==1.2.7
django-model-utils==1.4.0
django-mptt==0.6.1
django-polymorphic==0.5.6
django-qrcode==0.3
django-reversion==1.8.4
django-sekizai==0.7
django-simple-captcha==0.4.2
django-tagging==0.3.2
django-xmlrpc==0.1.5
easy-thumbnails==2.1
google-analytics==0.2
html5lib==0.999
logutils==0.3.3
psycopg2==2.5.3
pygeocoder==1.2.5
pygeoip==0.3.1
pyparsing==2.0.2
pytz==2014.7
requests==2.4.1
six==1.7.3
sqlparse==0.1.11